fix code readability
