+++ HomeForm.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Microsoft.Data.SqlClient;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -16,9 +17,40 @@ namespace Menu_Management
         {
             get => CategoryFlowPanel;
         }
+        private Image convertToImage(byte[] data)
+        {
+            byte[] imgData = data; //mảng chứa data của ảnh lấy từ CSDL (lúc này chính là byte)
+            Image img = null; //tạo biến Image để lưu ảnh sau khi convert
+            if (imgData != null) //khi ảnh được đọc lên khác null
+            {
+                using (MemoryStream ms = new MemoryStream(imgData)) //sử dụng đối tượng MemoryStream để lưu data ảnh (byte) được đọc vào luồng trực tiếp (bộ nhớ thay vì ổ cứng)
+                {
+                    img = Image.FromStream(ms); //chuyển đổi 1 luồng dữ liệu (byte) thành dạng ảnh thông qua FromStream và lưu vào biến img trước đó đã tạo
+                    return img;
+                }
+            }
+            else
+            {
+                return null;
+            }
+        }
         public HomeForm()
         {
             InitializeComponent();
+            using (SqlConnection sqlcon = new SqlConnection(DatabaseHelper.GetConnectionString()))
+            {
+                sqlcon.Open();
+                SqlCommand sqlcmd = new SqlCommand("SELECT * FROM Categories", sqlcon);
+                SqlDataReader reader = sqlcmd.ExecuteReader();
+                while (reader.Read())
+                {
+                    string categoryName = reader["CategoryName"].ToString();
+                    byte[] cateimgdata = reader["CategoryIMG"] as byte[];
+                    Image categoryImage = convertToImage(cateimgdata);
+                    UC_CategoryItem categoryItem = new UC_CategoryItem(categoryName, categoryImage);
+                    CategoryPanel.Controls.Add(categoryItem);
+                }
+            }
         }
     }
 }
